+++ crypto/sha-private.h
